+++ crates/provider/src/lib.rs
@@ -1,2 +1,3 @@
+mod helpers;
 mod native;
 mod shared;
